pyproject.toml
[project]
name = "oapython"
version = "0.1.0"
description = "OA办公系统纯后端API服务"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.3.0",
    "python-decouple>=3.8",
    "psycopg2-binary>=2.9.0",
    "pillow>=10.0.0",
    "django-filter>=23.0",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=0.10.0",
]
